requirements.txt
streamlit>=1.32.0
tensorflow>=2.13.0
keras>=3.0.0
numpy>=1.23.0
pillow>=9.0.0
